Phase one of TUI refactor
