+++ internal/tui/config/config_teatest_test.go
@@ -1,4 +1,4 @@
-package tui
+package config
 
 import (
 	"bytes"
