+++ internal/tui/progress/index_progress.go
@@ -1,4 +1,4 @@
-package tui
+package progress
 
 import (
 	"context"
