+++ internal/tui/progress/index_progress_teatest_test.go
@@ -1,4 +1,4 @@
-package tui
+package progress
 
 import (
 	"bytes"
