+++ internal/tui/progress/palette.go
@@ -1,0 +1,12 @@
+package progress
+
+import "github.com/Digital-Shane/title-tidy/internal/tui/components"
+
+var (
+	colorPrimary    = components.ColorPrimary
+	colorSecondary  = components.ColorSecondary
+	colorAccent     = components.ColorAccent
+	colorBackground = components.ColorBackground
+	colorSuccess    = components.ColorSuccess
+	colorError      = components.ColorError
+)
