+++ internal/tui/rename/palette.go
@@ -1,0 +1,13 @@
+package rename
+
+import "github.com/Digital-Shane/title-tidy/internal/tui/components"
+
+var (
+	colorPrimary    = components.ColorPrimary
+	colorSecondary  = components.ColorSecondary
+	colorAccent     = components.ColorAccent
+	colorBackground = components.ColorBackground
+	colorSuccess    = components.ColorSuccess
+	colorError      = components.ColorError
+	colorMuted      = components.ColorMuted
+)
