+++ internal/tui/rename/tui_teatest_test.go
@@ -1,4 +1,4 @@
-package tui
+package rename
 
 import (
 	"bytes"
