+++ internal/tui/undo/palette.go
@@ -1,0 +1,11 @@
+package undo
+
+import "github.com/Digital-Shane/title-tidy/internal/tui/components"
+
+var (
+	colorPrimary    = components.ColorPrimary
+	colorSecondary  = components.ColorSecondary
+	colorAccent     = components.ColorAccent
+	colorBackground = components.ColorBackground
+	colorMuted      = components.ColorMuted
+)
